Added emotion degradation
Emotions now degrade with time (depending on the updating of a person).
The equations are experimental at the moment and require further
testing. Within the application, pressing "Say Nothing" will force the
degradation as if time is passing. Currently, the assumed "update
frequency" of a person is 6 seconds. So every 6 seconds, someone will
say something to change an emotion, or their emotions will begin to
degrade.

Closes #10